Makes the monitored process names parametric
Moves the names of the processes that are being monitored to a global list.
Instead of checking them in a series of if statements, loop over the list to find the first one that is online.
This makes it easier to manipulate the list of checked processes for less expert users.